Link testObj.cpp
also had to change headers/update some files because of virtual weak assert lul
