+++ include/Screen/Game2DMgr.h
@@ -4,7 +4,7 @@
 #include "CNode.h"
 #include "JSystem/JKR/JKRDisposer.h"
 #include "JSystem/JUT/TColor.h"
-#include "Screen/Bases.h"
+#include "Screen/screenObj.h"
 #include "Screen/screenMgr.h"
 #include "types.h"
