+++ include/kh/FinalFloor.h
@@ -4,7 +4,7 @@
 #include "kh/khUtil.h"
 #include "og/Screen/DispMember.h"
 #include "og/Screen/callbackNodes.h"
-#include "Screen/Bases.h"
+#include "Screen/screenObj.h"
 #include "P2DScreen.h"
 
 struct JAISound;
