+++ src/plugProjectOgawaU/testScene.cpp
@@ -1,4 +1,4 @@
-#include "Screen/Bases.h"
+#include "Screen/screenObj.h"
 
 namespace Screen {
